Separated each page in different files.
