+++ lib/home.dart
@@ -1,0 +1,63 @@
+import 'package:cuetor/training.dart';
+import 'package:cuetor/usersettings.dart';
+import 'package:cuetor/userstats.dart';
+import 'package:flutter/material.dart';
+import 'main.dart';
+
+class CueTorHomePage extends StatefulWidget {
+  const CueTorHomePage({super.key, required this.title});
+
+  final String title;
+
+  @override
+  State<CueTorHomePage> createState() => _CueTorHomePageState();
+}
+
+class _CueTorHomePageState extends State<CueTorHomePage> {
+  int _selectedIndex = 0;
+
+  static const List<Widget> _widgetOptions = <Widget>[
+    TrainingPage(),
+    StatisticsPage(),
+    SettingsPage(),
+  ];
+
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Colors.blueGrey[900],
+        title: const Center(
+          child: Text('CueTor'),
+        ),
+      ),
+      body: _widgetOptions.elementAt(_selectedIndex),
+      bottomNavigationBar: BottomNavigationBar(
+        backgroundColor: Colors.blueGrey[900],
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            icon: Icon(Icons.school),
+            label: 'Training',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.analytics),
+            label: 'Statistics',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.settings),
+            label: 'Settings',
+          ),
+        ],
+        currentIndex: _selectedIndex,
+        selectedItemColor: Colors.blue,
+        onTap: _onItemTapped,
+      ),
+    );
+  }
+}
