+++ lib/training.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class TrainingPage extends StatelessWidget {
+  const TrainingPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text(
+        'Settings',
+        style: TextStyle(fontSize: 24),
+      ),
+    );
+  }
+}
