+++ lib/userstats.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class StatisticsPage extends StatelessWidget {
+  const StatisticsPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text(
+        'Player Statistics',
+        style: TextStyle(fontSize: 24),
+      ),
+    );
+  }
+}
