+++ pubspec.yaml
@@ -12,9 +12,11 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.2
-  camera: ^0.10.0+1
+  get: ^4.6.5
+  camera: ^0.10.5
+  image: ^4.0.17
   video_player: ^2.2.5
-  image_picker: ^0.8.4+11
+  image_picker: ^0.8.7+5
 
 dev_dependencies:
   flutter_test:
